disable some console
