+++ src/redux/redux.js
@@ -1,0 +1,27 @@
+import { createStore } from "redux";
+
+const globalState = {
+  user: "",
+};
+
+//Reducer
+const rootReducer = (state = globalState, action) => {
+  // console.log("aksi", action);
+
+  if (action.type === "SET_USER") {
+    return {
+      ...state,
+      user: action.userEmail,
+    };
+  }
+
+  return state;
+};
+
+//Store
+export const storeRedux = createStore(rootReducer);
+
+// Subscription
+storeRedux.subscribe(() => {
+  console.log("Redux : ", storeRedux.getState());
+});
